requirements.txt
fastapi
uvicorn
supabase
python-dotenv
requests
beautifulsoup4
certifi
pydantic
urllib3